Adds Entity Framework Core and PostgreSQL integration across services, along with initial database migrations for IdentityService

+++ Services/BookingService/BookingService_Domain/Entities/PaymentMethod.cs
@@ -6,8 +6,8 @@ public class PaymentMethod
     public string Name { get; set; } = string.Empty;
     public string Code { get; set; } = string.Empty;
     public string? IconUrl { get; set; } = string.Empty;
-    public float FeePercentage { get; set; }
-    public float FixedFee { get; set; }
+    public decimal FeePercentage { get; set; }
+    public decimal FixedFee { get; set; }
     public bool IsActive { get; set; } = true;
     
     public ICollection<Payment> Payments { get; set; } = new List<Payment>();
